+++ Source/sourcekitten/CompleteCommand.swift
@@ -66,7 +66,7 @@ struct CompleteCommand: CommandProtocol {
         }
 
         let request = Request.codeCompletionRequest(file: path, contents: contents,
-            offset: Int64(options.offset),
+            offset: options.offset,
             arguments: args)
         do {
             print(CodeCompletionItem.parse(response: try request.failableSend()))
